pequeños cambios

+++ src/main/java/com/example/restapi/model/Usuario.java
@@ -9,7 +9,7 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String nombre;
     private String apellido;
@@ -39,10 +39,10 @@ public class Usuario {
 
     // Getters y setters
     public Long getId() {
-        return (long) id;
+        return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
